rename test/util to test/utils
