+++ src/test/java/com/yelp/clientlib/utils/JsonTestUtils.java
@@ -1,4 +1,4 @@
-package com.yelp.clientlib.util;
+package com.yelp.clientlib.utils;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
